Update 10-scaling-up-road-to-the-top-part-3.ipynb
On [this point](https://youtu.be/p4ZZq0736Po?t=1104) of the L7 of Practical Deep Learning for Coders, you've explained that we pass the effective batch size to the constructor of the `GradientAccumulation` class. But the comment in the notebook doesn't match that fact and got a bit confusing to figure out what's the case actually.